refactor(core): remove redundant barrel files for better tree-shaking
Remove 9 redundant barrel files (index.ts) to improve tree-shaking:
- helpers/naming/index.ts
- helpers/path/index.ts
- types/ir/common/index.ts
- types/ir/endpoints/index.ts
- types/ir/models/index.ts
- types/ir/security/index.ts
- types/ir/tags/index.ts (converted to tags.ts)
- types/ir/metadata/index.ts (converted to metadata.ts)
- types/ir/servers/index.ts (converted to servers.ts)

All exports now use direct file imports instead of intermediate barrel
files, reducing the indirection and improving bundle optimization.

Verified with:
- pnpm typecheck ✓
- pnpm test ✓ (549 tests passed)
- pnpm lint ✓
