--- packages/core/src/transformer/helpers/naming/index.ts
@@ -1,10 +1,0 @@
-export { buildAdditionalPropertiesModelName } from "./build-additional-properties-model-name";
-export { buildInlineModelName } from "./build-inline-model-name";
-export { buildParameterModelName } from "./build-parameter-model-name";
-export { buildParameterSchemaModelName } from "./build-parameter-schema-model-name";
-export { buildRequestBodyModelName } from "./build-request-body-model-name";
-export { buildResponseModelName } from "./build-response-model-name";
-export { generateEnumName } from "./generate-enum-name";
-export { getModelName } from "./get-model-name";
-export { getMediaTypeSuffix } from "./media-type-suffix";
-export { pathToComponentBase } from "./path-to-component-base";
