--- packages/core/src/transformer/helpers/path/index.ts
@@ -1,9 +1,0 @@
-export { buildComponentSchemaPath } from "./build-component-schema-path";
-export { buildInlineSchemaPath } from "./build-inline-schema-path";
-export { buildReferencePath } from "./build-reference-path";
-export { parseCompositionPath } from "./parse-document-path";
-export { parseParameterPath } from "./parse-document-path";
-export { parseResponsePath } from "./parse-document-path";
-export { parseRequestBodyPath } from "./parse-document-path";
-export { parseAdditionalPropertiesPath } from "./parse-document-path";
-export { parseSchemaPath } from "./parse-document-path";
