--- packages/core/src/types/ir/common/index.ts
@@ -1,7 +1,0 @@
-/**
- * Common type definitions
- */
-
-export type { IRExtensions, IRExtensionValue } from "./extensions";
-export type { MimeType } from "./mime-type";
-export type { IRRef, IRScalarType, IRType } from "./type";
