--- packages/core/src/types/ir/models/index.ts
@@ -1,23 +1,0 @@
-/**
- * Model type definitions
- */
-
-export type {
-  IRAllOfModel,
-  IRAnyOfModel,
-  IRArrayModel,
-  IRDiscriminator,
-  IREnumModel,
-  IREnumValue,
-  IRMapModel,
-  IRObjectModel,
-  IRUnionModel,
-} from "./base";
-export type {
-  IRModel,
-  IRParameterModel,
-  IRRequestBodyModel,
-  IRResponseModel,
-} from "./operation";
-export type { IRParameterProperty, IRProperty } from "./property";
-export type { IRValidation } from "./validation";
